+++ ts/features/euCovidCert/utils/screenshot.ts
